requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
langgraph
faiss-cpu
pymupdf
tiktoken
streamlit
huggingface_hub 
tavily-python
